refactor: extract DndContext to providers, make layout SSR
Move DndContext and jotai draggable-tooltip logic from the params
layout into a client providers.tsx so the layout itself can render on
the server. Mark MapContainer as a client component since the layout no
longer carries the 'use client' boundary.
